+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
     -   There is a detailed list of the personnel that belongs to the vehicle.
     -   A list of patients that are in the vehicle is shown and the patients tab opens for details when clicking a patient.
     -   Vehicles can be deleted or moved to the map.
+    -   Trainers can see the current occupation of a vehicle and cancel it.
 -   Simulated regions are prefixed with "\[Simuliert\]" in the request target selection for the requests behavior.
 
 ### Changed